rename: theme type, variable 이동
