+++ src/styles/theme.ts
@@ -1,4 +1,4 @@
-import {TypeOfTheme} from 'types/theme';
+import {TypeOfTheme} from 'styles/types';
 import {palette} from 'styles';
 
 export const theme: TypeOfTheme = {
