+++ src/styles/types/index.ts
